backend/requirements.txt
fastapi~=0.115.6
llama-index-core~=0.12.5
SQLAlchemy~=2.0.36
scikit-learn~=1.6.0
pydantic~=2.10.3
mistralai~=1.2.5